Move `resolve_invoc` from `syntax` to `resolve`.
